Reseñas arregladas

+++ src/main/java/businessLogic/BLFacade.java
@@ -85,5 +85,5 @@ public interface BLFacade  {
 	@WebMethod public List<Reportes> getReportesEnviados(String sellerMail);
 	@WebMethod public List<Reportes> getReportesRecibidos();
 	
-	@WebMethod public boolean hayRese(String eVend, String eComp, float idP);
+	@WebMethod public boolean hayRese(String eVend, String eComp, long idP);
  }
